# storage/docs/openapi.yaml
openapi: 3.0.3
info:
  description: "This is the Collectio Server API"
  version: 1.0.0
  title: Collectio
  termsOfService: https://collectioapp.com/
  contact:
    email: support@collectioapp.com
  license:
    name: MIT
    url: https://mit-license.org
servers:
  - url: https://collectioapp.com/api/v1
  - url: https://beta.collectioapp.com/api/v1
security:
  - basicAuth: []
  - apiKey: []
...
paths:
...
  /items:
    get:
      tags:
        - items
      summary: List items
      description: ""
      operationId: listItems
      parameters:
        - name: collectionId
          in: query
          description: Filter items by collection id
          required: false
          explode: true
          schema:
            type: integer
            format: int64
        - name: itemId
          in: query
          description: Filter items by item id
          required: false
          explode: true
          schema:
            type: integer
            format: int64
        - name: others
          in: query
          description: Filter by other users
          required: false
          schema:
            type: boolean
            default: false
        - name: subcategoryId
          in: query
          description: Filter items by subcategory id
          required: false
          explode: true
          schema:
            type: integer
            format: int64
        - name: withChildren
          in: query
          description: Return collections with childrens
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/InternalError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ForbiddenBasicAuth"
        "200":
          description: items
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ListItems"
    put:
      tags:
        - items
      summary: New item
      description: ""
      operationId: newItem
      requestBody:
        description: ""
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/Item"
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/InternalError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ForbiddenBasicAuth"
        "200":
          description: item
          content:
            application/json:
              schema:
                $ref: "#/components/responses/NewItem"
    post:
      tags:
        - items
      summary: Edit item
      description: ""
      operationId: editItem
      requestBody:
        description: ""
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/Item"
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/InternalError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ForbiddenBasicAuth"
        "200":
          description: item
          content:
            application/json:
              schema:
                $ref: "#/components/responses/EditItem"
    delete:
      tags:
        - items
      summary: Delete item
      description: ""
      operationId: deleteItem
      parameters:
        - name: ServerId
          in: path
          description: Item to delete
          required: true
          explode: true
          schema:
            type: integer
            format: int64
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/InternalError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ForbiddenBasicAuth"
        "200":
          description: item
          content:
            application/json:
              schema:
                $ref: "#/components/responses/DeleteItem"
  /items/images:
    get:
      tags:
        - itemImages
      summary: List item images
      description: ""
      operationId: listItemImages
      parameters:
        - name: ItemId
          in: query
          description: Filter item images by item id
          required: true
          explode: true
          schema:
            type: integer
            format: int64
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/InternalError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ForbiddenBasicAuth"
        "200":
          description: item images
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ListItemImages"
    put:
      tags:
        - itemImages
      summary: New item image
      description: ""
      operationId: newItemImage
      requestBody:
        description: ""
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ItemImage"
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/InternalError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ForbiddenBasicAuth"
        "200":
          description: item image
          content:
            application/json:
              schema:
                $ref: "#/components/responses/NewItemImage"
    delete:
      tags:
        - itemImages
      summary: Delete item image
      description: ""
      operationId: deleteItemImage
      parameters:
        - name: ServerId
          in: path
          description: Item image to delete
          required: true
          explode: true
          schema:
            type: integer
            format: int64
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/InternalError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ForbiddenBasicAuth"
        "200":
          description: item
          content:
            application/json:
              schema:
                $ref: "#/components/responses/DeleteItemImage"
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Token
  responses:
    ForbiddenBasicAuth:
      $ref: "#/components/schemas/ResponseWS"
      # Code 403, Status: 'KO'
...
    InternalError:
      $ref: "#/components/schemas/ResponseWS"
      # Code 500, Status: 'KO'
...
  schemas:
    ResponseWS:
      type: object
      properties:
        status:
          type: string
          minLength: 2
          maxLength: 2
        code:
          type: integer
          format: int64
        message:
          type: string
        data:
          nullable: true
          oneOf:
            - type: array
              items:
                $ref: "#/components/schemas/Category"
            - type: array
              items:
                $ref: "#/components/schemas/Subcategory"
            - type: array
              items:
                $ref: "#/components/schemas/Collection"
            - type: array
              items:
                $ref: "#/components/schemas/Item"
            - type: array
              items:
                $ref: "#/components/schemas/ItemImage"
            - type: array
              items:
                $ref: "#/components/schemas/Apikey"
            - $ref: "#/components/schemas/User"
    Category:
      type: object
      properties:
        Id:
          type: integer
          format: int64
        Image:
          type: string
        Spanish:
          type: string
        English:
          type: string
        Catalan:
          type: string
        Basque:
          type: string
    Subcategory:
      type: object
      properties:
        Id:
          type: integer
          format: int64
        CategoryId:
          type: integer
          format: int64
        Image:
          type: string
        Spanish:
          type: string
        English:
          type: string
        Catalan:
          type: string
        Basque:
          type: string
    Collection:
      type: object
      properties:
        ServerId:
          type: integer
          format: int64
        UserId:
          type: integer
          format: int64
        CategoryId:
          type: integer
          format: int64
        Name:
          type: string
        Description:
          type: string
        Private:
          type: boolean
        CreatedAt:
          type: string
          format: date-time
        UpdatedAt:
          type: string
          format: date-time
        Image:
          type: string
        Items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
    Item:
      type: object
      properties:
        ServerId:
          type: integer
          format: int64
        CollectionId:
          type: integer
          format: int64
        SubcategoryId:
          type: integer
          format: int64
        Name:
          type: string
        Description:
          type: string
        Private:
          type: boolean
        CreatedAt:
          type: string
          format: date-time
        UpdatedAt:
          type: string
          format: date-time
        Images:
          type: array
          items:
            $ref: "#/components/schemas/ItemImage"
    ItemImage:
      type: object
      properties:
        ServerId:
          type: integer
          format: int64
        ItemId:
          type: integer
          format: int64
        Image:
          type: string
    User:
      type: object
      properties:
        Nickname:
          type: string
        Mail:
          type: string
        Image:
          type: string
        Apple:
          type: boolean
        Google:
          type: boolean
    Apikey:
      type: object
      properties:
        Token:
          type: string
        CreatedAt:
          type: string
          format: date-time
        UsedAt:
          type: string
          format: date-time
        Device:
          type: string
        UserDeviceName:
          type: string